+++ src/main/java/opennote/file/FileController.java
@@ -1,4 +1,4 @@
-package opennote.File;
+package opennote.file;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
