+++ src/main/java/opennote/file/FileHandler.java
@@ -1,4 +1,4 @@
-package opennote.File;
+package opennote.file;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.multipart.MultipartFile;
