+++ src/main/java/opennote/file/S3Config.java
@@ -1,4 +1,4 @@
-package opennote.File;
+package opennote.file;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
