+++ src/main/java/opennote/folder/Folder.java
@@ -1,9 +1,10 @@
-package opennote.Note;
+package opennote.folder;
 
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import opennote.Folder.Folder;
-import opennote.User.User;
+import opennote.note.Note;
+import opennote.user.User;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -13,8 +14,8 @@ import java.util.Date;
 import java.util.List;
 
 @Entity
-@Table(name = "notes")
-public class Note {
+@Table( name = "folders" )
+public class Folder {
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "uuid")
@@ -23,37 +24,34 @@ public class Note {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
-    @JsonIgnoreProperties(value = {"hibernateLazyInitializer", "notes"})
-    private User user;
-
-    @ManyToMany(fetch = FetchType.LAZY, mappedBy = "notes")
     @JsonBackReference
-    private List<Folder> folders = new ArrayList<>();
+    private User user;
+    @ManyToMany(fetch = FetchType.LAZY)
+    @JoinTable(
+            name = "folder_notes",
+            joinColumns = @JoinColumn(name = "folder_id"),
+            inverseJoinColumns = @JoinColumn(name = "note_id"))
+    @JsonManagedReference
+    private List<Note> notes = new ArrayList<>();
     private String title;
-    private String url;
-    private boolean isPublic;
     @CreationTimestamp
     private Date createdAt;
+
     @UpdateTimestamp
     private Date updatedAt;
 
-    public Note(){}
+    public Folder(){}
 
-    public Note(String id, User user, String title, String url, boolean isPublic, Date createdAt, Date updatedAt) {
+    public Folder(String id, User user, String title, Date createdAt, Date updatedAt) {
         this.id = id;
         this.user = user;
         this.title = title;
-        this.url = url;
-        this.isPublic = isPublic;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
     }
 
     public String getId() {
         return id;
-    }
-    public String getTitle() {
-        return title;
     }
 
     public User getUser() {
@@ -64,24 +62,12 @@ public class Note {
         this.user = user;
     }
 
+    public String getTitle() {
+        return title;
+    }
+
     public void setTitle(String title) {
         this.title = title;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public boolean isPublic() {
-        return isPublic;
-    }
-
-    public void setPublic(boolean aPublic) {
-        isPublic = aPublic;
     }
 
     public Date getCreatedAt() {
@@ -98,5 +84,13 @@ public class Note {
 
     public void setUpdatedAt(Date updatedAt) {
         this.updatedAt = updatedAt;
+    }
+
+    public void addNote(Note note){
+        notes.add(note);
+    }
+
+    public void removeNote(Note note){
+        notes.remove(note);
     }
 }
