+++ src/main/java/opennote/folder/FolderNotFoundException.java
@@ -1,4 +1,4 @@
-package opennote.Folder;
+package opennote.folder;
 
 public class FolderNotFoundException extends RuntimeException{
     public FolderNotFoundException(String id) {
