+++ src/main/java/opennote/folder/FolderRepository.java
@@ -1,8 +1,7 @@
-package opennote.Folder;
+package opennote.folder;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
