+++ src/main/java/opennote/note/NoteNotFoundException.java
@@ -1,4 +1,4 @@
-package opennote.Note;
+package opennote.note;
 
 public class NoteNotFoundException extends RuntimeException{
     public NoteNotFoundException(String id) { super("Could not find note " + id);}
