+++ src/main/java/opennote/note/NoteRepository.java
@@ -1,4 +1,4 @@
-package opennote.Note;
+package opennote.note;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
