+++ src/main/java/opennote/user/Role.java
@@ -1,4 +1,4 @@
-package opennote.User;
+package opennote.user;
 
 public enum Role {
     USER,
