+++ src/main/java/opennote/user/UserRepository.java
@@ -1,10 +1,8 @@
-package opennote.User;
+package opennote.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
