+++ src/main/java/opennote/user/UserService.java
@@ -1,6 +1,5 @@
-package opennote.User;
+package opennote.user;
 
-import opennote.Folder.Request.NewFolderRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
